benchmarks made easier to understand
