+++ benchmarks/lib/CountAndLengthBench.php
@@ -15,27 +15,11 @@ class CountAndLengthBench
         (new Stringy('Hello world'))->length();
     }
 
-    public function bench_indexOf_Str() {
-        (new Str('Hello world'))->indexOf('l');
-    }
-
-    public function bench_indexOf_Stringy() {
-        (new Stringy('Hello world'))->indexOf('l');
-    }
-
-    public function bench_indexOfLast_Str() {
-        (new Str('Hello world'))->indexOfLast('l');
-    }
-
-    public function bench_indexOfLast_Stringy() {
-        (new Stringy('Hello world'))->indexOfLast('l');
-    }
-
-    public function bench_count_Str() {
+    public function bench_count_substr_Str() {
         (new Str('HelLo wOrld'))->countSubstr('l', false);
     }
 
-    public function bench_count_Stringy() {
+    public function bench_count_substr_Stringy() {
         (new Stringy('HelLo wOrld'))->countSubstr('l', false);
     }
 }
