+++ benchmarks/lib/IsOrHasUpperCaseBench.php
@@ -5,7 +5,7 @@ namespace Benchmark;
 use Str\Str;
 use Stringy\Stringy;
 
-class CaseQuestionsBench
+class IsOrHasUpperCaseBench
 {
     public function bench_is_upper_case_Str() {
         (new Str(' hello world '))->isUpperCase();
@@ -15,27 +15,11 @@ class CaseQuestionsBench
         (new Stringy(' hello world '))->isUpperCase();
     }
 
-    public function bench_is_lower_case_Str() {
-        (new Str(' hello world '))->isLowerCase();
-    }
-
-    public function bench_is_lower_case_Stringy() {
-        (new Stringy(' hello world '))->isLowerCase();
-    }
-
     public function bench_has_upper_case_Str() {
         (new Str(' hello world '))->hasUpperCase();
     }
 
     public function bench_has_upper_case_Stringy() {
         (new Stringy(' hello world '))->hasUpperCase();
-    }
-
-    public function bench_has_lower_case_Str() {
-        (new Str(' hello world '))->hasLowerCase();
-    }
-
-    public function bench_has_lower_case_Stringy() {
-        (new Stringy(' hello world '))->hasLowerCase();
     }
 }
