--- benchmarks/lib/SubstrBench.php
@@ -1,81 +1,0 @@
-<?php
-
-namespace Benchmark;
-
-use Str\Str;
-use Stringy\Stringy;
-
-class SubstrBench
-{
-    public function bench_slice_Str() {
-        (new Str(' hello world '))->slice(2);
-    }
-
-    public function bench_slice_Stringy() {
-        (new Stringy(' hello world '))->slice(2);
-    }
-
-    public function bench_longest_common_substring_Str() {
-        (new Str(' hello world '))->longestCommonSubstring(' helloworld');
-    }
-
-    public function bench_longest_common_substring_Stringy() {
-        (new Stringy(' hello world '))->longestCommonSubstring(' helloworld');
-    }
-
-    public function bench_longest_common_suffix_Str() {
-        (new Str(' hello world '))->longestCommonSuffix(' hello');
-    }
-
-    public function bench_longest_common_suffix_Stringy() {
-        (new Stringy(' hello world '))->longestCommonSuffix(' hello');
-    }
-
-    public function bench_longest_common_prefix_Str() {
-        (new Str(' hello world '))->longestCommonPrefix(' hello');
-    }
-
-    public function bench_longest_common_prefix_Stringy() {
-        (new Stringy(' hello world '))->longestCommonPrefix(' hello');
-    }
-
-    public function bench_between_Str() {
-        (new Str(' hello world '))->between(2, 4);
-    }
-
-    public function bench_between_Stringy() {
-        (new Stringy(' hello world '))->between(2, 4);
-    }
-
-    public function bench_last_Str() {
-        (new Str(' hello world '))->last(2);
-    }
-
-    public function bench_last_Stringy() {
-        (new Stringy(' hello world '))->last(2);
-    }
-
-    public function bench_first_Str() {
-        (new Str(' hello world '))->first(2);
-    }
-
-    public function bench_first_Stringy() {
-        (new Stringy(' hello world '))->first(2);
-    }
-
-    public function bench_at_Str() {
-        (new Str(' hello world '))->at(2);
-    }
-
-    public function bench_at_Stringy() {
-        (new Stringy(' hello world '))->at(2);
-    }
-
-    public function bench_substr_Str() {
-        (new Str(' hello world '))->substr(2);
-    }
-
-    public function bench_substr_Stringy() {
-        (new Stringy(' hello world '))->substr(2);
-    }
-}
